HeaderからロジックをAppに移した
