+++ frontend/src/components/Header.tsx
@@ -1,30 +1,22 @@
 import React from 'react';
-import { NavLink, useHistory } from 'react-router-dom';
 import { useCookies } from 'react-cookie';
+import { NavLink, useHistory } from 'react-router-dom';
 import ExitToAppIcon from '@material-ui/icons/ExitToApp';
-import verifyToken from 'commons/auth/verifyToken';
 import styles from './Header.module.css';
 
-const Header: React.FC = () => {
-  const [login, setLogin] = React.useState(false);
+type Props = {
+  isLogin: boolean;
+  setIsLogin: (param: boolean) => void;
+};
+
+const Header: React.FC<Props> = ({ isLogin, setIsLogin }) => {
   const history = useHistory();
-  const [cookie] = useCookies();
   const removeCookie = useCookies()[2];
   const logout = () => {
     removeCookie('calendarJWT');
-    setLogin(false);
+    setIsLogin(false);
     history.push('/login');
   };
-
-  React.useEffect(() => {
-    const checkAlreadyAuthenticated = async () => {
-      if (cookie.hasOwnProperty('calendarJWT')) {
-        const isAuthenticated = await verifyToken(cookie.calendarJWT);
-        isAuthenticated && setLogin(isAuthenticated);
-      }
-    };
-    checkAlreadyAuthenticated();
-  }, [cookie]);
 
   const HeaderLink = () => {
     return (
@@ -52,7 +44,7 @@ const Header: React.FC = () => {
       <NavLink to="/" className={styles.header_logo_nav}>
         Scrum Calendar
       </NavLink>
-      {login && <HeaderLink />}
+      {isLogin && <HeaderLink />}
     </header>
   );
 };
